+++ src/main/java/com/example/demo/domain/apply/controller/ApplyController.java
@@ -1,9 +1,7 @@
 package com.example.demo.domain.apply.controller;
 
-import com.example.demo.domain.apply.DTO.ApplyResponseDTO;
+import com.example.demo.domain.apply.dto.ApplyResponseDTO;
 import com.example.demo.domain.apply.service.ApplyService;
-import com.example.demo.domain.user.entity.User;
-import com.example.demo.domain.user.service.UserService;
 import com.example.demo.global.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
@@ -26,7 +24,6 @@ public class ApplyController {
             @PathVariable Long eventId,
             @PathVariable Long sessionId
     ) {
-        ApplyResponseDTO responseDTO = applyService.applyToSession(eventId, sessionId);
-        return ApiResponse.onSuccess(_OK, responseDTO);
+        return ApiResponse.onSuccess(_OK, applyService.applyToSession(eventId, sessionId));
     }
 }
