+++ src/main/java/com/example/demo/domain/apply/service/ApplyService.java
@@ -1,8 +1,7 @@
 package com.example.demo.domain.apply.service;
 
-import com.example.demo.domain.apply.DTO.ApplyResponseDTO;
+import com.example.demo.domain.apply.dto.ApplyResponseDTO;
 import com.example.demo.domain.event.entity.Event;
-import com.example.demo.domain.user.entity.User;
 
 public interface ApplyService {
