+++ src/main/java/com/example/demo/domain/event/controller/BuyerEventController.java
@@ -1,6 +1,6 @@
 package com.example.demo.domain.event.controller;
 
-import com.example.demo.domain.event.dto.response.BuyerEventInfoDTO;
+import com.example.demo.domain.event.dto.response.BuyerEventDetailDTO;
 import com.example.demo.domain.event.service.BuyerEventService;
 import com.example.demo.global.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
@@ -21,7 +21,7 @@ public class BuyerEventController {
 
     @Operation(summary = "구매자 - 공연 상세 조회")
     @GetMapping("{eventId}")
-    public ApiResponse<BuyerEventInfoDTO> getEventDetail(@PathVariable Long eventId) {
+    public ApiResponse<BuyerEventDetailDTO> getEventDetailForBuyer(@PathVariable Long eventId) {
         return ApiResponse.onSuccess(_OK, buyerEventService.getEventDetailForBuyer(eventId));
     }
 }
