+++ src/main/java/com/example/demo/domain/event/controller/OrganizerEventController.java
@@ -1,7 +1,7 @@
 package com.example.demo.domain.event.controller;
 
 import com.example.demo.domain.event.dto.request.EventUploadDTO;
-import com.example.demo.domain.event.dto.response.OrganizerEventInfoDTO;
+import com.example.demo.domain.event.dto.response.OrganizerEventDetailDTO;
 import com.example.demo.domain.event.dto.response.ResponseDTO;
 import com.example.demo.domain.event.dto.response.OrganizerSessionInfoDTO;
 import com.example.demo.domain.event.service.OrganizerEventService;
@@ -26,8 +26,8 @@ public class OrganizerEventController {
 
     @Operation(summary = "개최자 - 공연 상세 조회")
     @GetMapping("{eventId}")
-    public ApiResponse<OrganizerEventInfoDTO> getEventInfoForOrganizer(@PathVariable Long eventId) {
-        return ApiResponse.onSuccess(_OK, organizerEventService.getEventInfoForOrganizer(eventId));
+    public ApiResponse<OrganizerEventDetailDTO> getEventDetailForOrganizer(@PathVariable Long eventId) {
+        return ApiResponse.onSuccess(_OK, organizerEventService.getEventDetailForOrganizer(eventId));
     }
 
     @Operation(summary = "개최자 - 회차 상세 조회")
@@ -45,9 +45,9 @@ public class OrganizerEventController {
             @RequestPart(name = "request") EventUploadDTO request,
             @RequestPart(name = "banner") MultipartFile banner,
             @RequestPart(name = "poster") MultipartFile poster,
-            @RequestPart(name = "photocardList") List<MultipartFile> photocardList
+            @RequestPart(name = "photocardList") List<MultipartFile> photoCardList
     ) {
-        return ApiResponse.onSuccess(_OK, organizerEventService.uploadEvent(request, banner, poster, photocardList));
+        return ApiResponse.onSuccess(_OK, organizerEventService.uploadEvent(request, banner, poster, photoCardList));
     }
 
 }
