+++ src/main/java/com/example/demo/domain/event/dto/response/BuyerEventDetailDTO.java
@@ -2,7 +2,6 @@ package com.example.demo.domain.event.dto.response;
 
 import com.example.demo.domain.event.enums.AgeLimit;
 import com.example.demo.domain.event.enums.EventStatus;
-import com.example.demo.domain.metadata.dto.PhotoCardInfoDTO;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,23 +16,27 @@ import java.util.List;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class OrganizerEventInfoDTO {
-
+public class BuyerEventDetailDTO {
     private Long eventId;
     private String title;
-    private String posterUrl;
+    private String description;
     private String location;
-    private LocalDate startDate;
-    private LocalDate endDate;
-    private LocalTime startTime;
-    private LocalTime endTime;
     private AgeLimit ageLimit;
     private int priceKrw;
+
     private LocalDateTime applyStart;
     private LocalDateTime applyEnd;
+
+    private LocalDate startDate;
+    private LocalDate endDate;
+
+    private LocalTime startTime;
+    private LocalTime endTime;
+
+    private String posterUrl;
     private int capacity;
+
     private EventStatus eventStatus;
-    private List<Long> sessionIds;
-    private String description;
-    private List<PhotoCardInfoDTO> photoCardList;
+
+    private List<BuyerSessionInfoDTO> sessionList;
 }
