+++ src/main/java/com/example/demo/domain/event/dto/response/OrganizerEventDetailDTO.java
@@ -2,6 +2,7 @@ package com.example.demo.domain.event.dto.response;
 
 import com.example.demo.domain.event.enums.AgeLimit;
 import com.example.demo.domain.event.enums.EventStatus;
+import com.example.demo.domain.metadata.dto.PhotoCardInfoDTO;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -16,27 +17,23 @@ import java.util.List;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class BuyerEventInfoDTO {
+public class OrganizerEventDetailDTO {
+
     private Long eventId;
     private String title;
-    private String description;
+    private String posterUrl;
     private String location;
-    private AgeLimit ageLimit;
-    private int priceKrw;
-
-    private LocalDateTime applyStart;
-    private LocalDateTime applyEnd;
-
     private LocalDate startDate;
     private LocalDate endDate;
-
     private LocalTime startTime;
     private LocalTime endTime;
-
-    private String posterUrl;
+    private AgeLimit ageLimit;
+    private int priceKrw;
+    private LocalDateTime applyStart;
+    private LocalDateTime applyEnd;
     private int capacity;
-
     private EventStatus eventStatus;
-
-    private List<BuyerSessionInfoDTO> sessionList;
+    private List<Long> sessionIds;
+    private String description;
+    private List<PhotoCardInfoDTO> photoCardList;
 }
