+++ src/main/java/com/example/demo/domain/event/repository/SessionRepository.java
@@ -13,4 +13,5 @@ import java.util.Optional;
 public interface SessionRepository extends JpaRepository<Session, Long> {
     Optional<Session> findById(Long id);
 
+    Optional<Session> findByIdAndEventId(Long id, Long eventId);
 }
