+++ src/main/java/com/example/demo/domain/event/service/OrganizerEventService.java
@@ -1,7 +1,7 @@
 package com.example.demo.domain.event.service;
 
 import com.example.demo.domain.event.dto.request.EventUploadDTO;
-import com.example.demo.domain.event.dto.response.OrganizerEventInfoDTO;
+import com.example.demo.domain.event.dto.response.OrganizerEventDetailDTO;
 import com.example.demo.domain.event.dto.response.ResponseDTO;
 import com.example.demo.domain.event.dto.response.OrganizerSessionInfoDTO;
 import org.springframework.web.multipart.MultipartFile;
@@ -10,7 +10,7 @@ import java.util.List;
 
 public interface OrganizerEventService {
 
-    OrganizerEventInfoDTO getEventInfoForOrganizer(Long eventId);
+    OrganizerEventDetailDTO getEventDetailForOrganizer(Long eventId);
 
     OrganizerSessionInfoDTO getSessionInfoForOrganizer(Long eventId, Long sessionId);
