+++ src/main/java/com/example/demo/domain/event/service/impl/BuyerEventServiceImpl.java
@@ -4,13 +4,11 @@ import com.example.demo.domain.apply.entity.Apply;
 import com.example.demo.domain.apply.enums.ApplyStatus;
 import com.example.demo.domain.apply.repository.ApplyRepository;
 import com.example.demo.domain.event.converter.EventConverter;
-import com.example.demo.domain.event.dto.response.BuyerEventInfoDTO;
+import com.example.demo.domain.event.dto.response.BuyerEventDetailDTO;
 import com.example.demo.domain.event.dto.response.BuyerSessionInfoDTO;
 import com.example.demo.domain.event.entity.Event;
 import com.example.demo.domain.event.entity.Session;
-import com.example.demo.domain.event.enums.EventStatus;
 import com.example.demo.domain.event.repository.EventRepository;
-import com.example.demo.domain.event.repository.SessionRepository;
 import com.example.demo.domain.event.service.BuyerEventService;
 import com.example.demo.domain.ticket.entity.Ticket;
 import com.example.demo.domain.ticket.repository.TicketRepository;
@@ -22,9 +20,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
+import static com.example.demo.domain.event.converter.SessionConverter.toBuyerSessionInfoDTO;
 
 @Service
 @RequiredArgsConstructor
@@ -32,13 +33,12 @@ import java.util.Optional;
 public class BuyerEventServiceImpl implements BuyerEventService {
 
     private final EventRepository eventRepository;
-    private final SessionRepository sessionRepository;
     private final ApplyRepository applyRepository;
     private final TicketRepository ticketRepository;
     private final UserService userService;
 
     @Override
-    public BuyerEventInfoDTO getEventDetailForBuyer(Long eventId) {
+    public BuyerEventDetailDTO getEventDetailForBuyer(Long eventId) {
         User user = userService.getCurrentUser();  // 현재 로그인한 구매자
 
         Event event = eventRepository.findById(eventId)
@@ -46,31 +46,50 @@ public class BuyerEventServiceImpl implements BuyerEventService {
 
         List<Session> sessions = event.getSessions();
 
+        List<Long> sessionIds = sessions.stream()
+                .map(Session::getId)
+                .toList();
+
+        List<Apply> applyList = applyRepository.findByUserIdAndSessionIdIn(user.getId(), sessionIds);
+        Map<Long, Apply> applyMap = applyList.stream()
+                .collect(Collectors.toMap(apply -> apply.getSession().getId(), Function.identity()));
+
+        List<Ticket> ticketList = ticketRepository.findByUserIdAndSessionIdIn(user.getId(), sessionIds);
+        Map<Long, Ticket> ticketMap = ticketList.stream()
+                .collect(Collectors.toMap(ticket -> ticket.getSession().getId(), Function.identity()));
+
         List<BuyerSessionInfoDTO> sessionDTOs = sessions.stream().map(session -> {
-            // 해당 세션에 대해 유저의 응모 정보 가져오기
-            Apply apply = applyRepository.findBySessionAndUser(session, user).orElse(null);
+            Apply apply = applyMap.get(session.getId());
             ApplyStatus applyStatus = apply != null ? apply.getApplyStatus() : null;
 
-            // 해당 세션에 대해 유저의 티켓 정보 가져오기
-            Ticket ticket = ticketRepository.findByUserAndSession(user, session).orElse(null);
+            Ticket ticket = ticketMap.get(session.getId());
             Long ticketId = ticket != null ? ticket.getId() : null;
 
-            // 잔여 티켓 계산
-            int paidCount = session.getTicketList().size();
-            boolean buyable = event.getEventStatus() == EventStatus.TICKETED
-                    && (event.getCapacity() - paidCount > 0)
-                    && applyStatus != ApplyStatus.PAID;
+            boolean buyable = validateBuyer(session, user, apply);
 
-            return BuyerSessionInfoDTO.builder()
-                    .sessionId(session.getId())
-                    .date(session.getDate())
-                    .paidCount(paidCount)
-                    .applyStatus(applyStatus)
-                    .ticketId(ticketId)
-                    .buyable(buyable)
-                    .build();
+            return toBuyerSessionInfoDTO(session, applyStatus, ticketId, buyable);
         }).toList();
 
         return EventConverter.toBuyerEventInfoDTO(event, sessionDTOs);
+    }
+
+    private boolean validateBuyer(Session session, User user, Apply apply) {
+        Event event = session.getEvent();
+
+        if (!user.isEligibleFor(event.getAgeLimit()) || !session.getIsBuyable()) {
+            return false;
+        }
+
+        switch (event.getEventStatus()) {
+            case APPLY_CLOSED:
+                return apply != null && apply.getApplyStatus() == ApplyStatus.SELECTED;
+
+            case TICKETED:
+            case IN_PROGRESS:
+                return true;
+
+            default:
+                return false;
+        }
     }
 }
