+++ src/main/java/com/example/demo/domain/event/service/impl/OrganizerEventServiceImpl.java
@@ -1,28 +1,27 @@
 package com.example.demo.domain.event.service.impl;
 
+import com.example.demo.domain.event.converter.SessionConverter;
 import com.example.demo.domain.event.dto.request.EventUploadDTO;
 import com.example.demo.domain.event.dto.response.ResponseDTO;
 import com.example.demo.domain.event.dto.response.OrganizerSessionInfoDTO;
 import com.example.demo.domain.event.entity.Event;
 import com.example.demo.domain.event.entity.Session;
-import com.example.demo.domain.event.enums.EventStatus;
 import com.example.demo.domain.event.repository.EventRepository;
-import com.example.demo.domain.event.dto.response.OrganizerEventInfoDTO;
+import com.example.demo.domain.event.dto.response.OrganizerEventDetailDTO;
 import com.example.demo.domain.event.repository.SessionRepository;
 import com.example.demo.domain.event.service.OrganizerEventService;
+import com.example.demo.domain.event.service.SessionService;
 import com.example.demo.domain.metadata.dto.PhotoCardInfoDTO;
-import com.example.demo.domain.metadata.entity.PhotoCard;
 import com.example.demo.domain.metadata.service.PhotoCardService;
 import com.example.demo.domain.user.entity.User;
 import com.example.demo.domain.user.service.UserService;
 import com.example.demo.global.base.Constants;
-import com.example.demo.global.infra.image.S3UploadService;
 import com.example.demo.global.infra.blockchain.service.DketNFTService;
+import com.example.demo.global.infra.image.S3UploadService;
 import com.example.demo.global.infra.blockchain.service.ExchangeService;
 import com.example.demo.global.infra.ipfs.PinataService;
 import com.example.demo.global.infra.scheduling.SchedulingService;
 import com.example.demo.global.infra.scheduling.jobs.event.OpenApplyJob;
-import com.example.demo.global.infra.scheduling.jobs.session.ClosePaymentJob;
 import com.example.demo.global.response.exception.CustomException;
 import com.example.demo.global.response.status.ErrorStatus;
 import lombok.RequiredArgsConstructor;
@@ -32,11 +31,10 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static com.example.demo.domain.event.converter.EventConverter.*;
 import static com.example.demo.domain.metadata.converter.PhotoCardConverter.toPhotoCardInfoDTO;
@@ -55,24 +53,25 @@ public class OrganizerEventServiceImpl implements OrganizerEventService {
     private final ExchangeService exchangeService;
     private final PhotoCardService photoCardService;
     private final PinataService pinataService;
+    private final SessionService sessionService;
 
     @Override
-    public OrganizerEventInfoDTO getEventInfoForOrganizer(Long eventId) {
+    public OrganizerEventDetailDTO getEventDetailForOrganizer(Long eventId) {
         User user = userService.getCurrentUser();
 
         Event event = eventRepository.findById(eventId)
                 .orElseThrow(() -> new CustomException(ErrorStatus.EVENT_NOT_FOUND));
 
-        if (!event.getOrganizer().equals(user))
-            throw new CustomException(ErrorStatus.EVENT_ORGANIZER_MISMATCH);
+        validateOrganizer(event, user);
 
-        List<PhotoCardInfoDTO> photoCardInfoDTOList = new ArrayList<>();
-        for (PhotoCard photoCard : event.getPhotoCards()) {
-            String url = pinataService.cidToHttp(photoCard.getCid());
-            photoCardInfoDTOList.add(toPhotoCardInfoDTO(photoCard, url));
-        }
+        List<PhotoCardInfoDTO> photoCardInfoDTOList = event.getPhotoCards().stream()
+                .map(photoCard -> {
+                    String url = pinataService.cidToHttp(photoCard.getCid());
+                    return toPhotoCardInfoDTO(photoCard, url);
+                })
+                .collect(Collectors.toList());
 
-        return toEventInfoDTO(event, photoCardInfoDTOList);
+        return toOrganizerEventInfoDTO(event, photoCardInfoDTOList);
     }
 
     @Override
@@ -82,20 +81,16 @@ public class OrganizerEventServiceImpl implements OrganizerEventService {
         Event event = eventRepository.findById(eventId)
                 .orElseThrow(() -> new CustomException(ErrorStatus.EVENT_NOT_FOUND));
 
-        Session session = sessionRepository.findById(sessionId)
+        Session session = sessionRepository.findByIdAndEventId(sessionId, eventId)
                 .orElseThrow(() -> new CustomException(ErrorStatus.SESSION_NOT_FOUND));
 
-        if (!event.getOrganizer().equals(user))
-            throw new CustomException(ErrorStatus.EVENT_ORGANIZER_MISMATCH);
-
-        if (!session.getEvent().equals(event))
-            throw new CustomException(ErrorStatus.EVENT_SESSION_MISMATCH);
+        validateOrganizer(event, user);
 
         int attendeeCount = (int) session.getTicketList().stream()
                 .filter(ticket -> ticket.getEnteredAt() != null)
                 .count();
 
-        return toSessionInfoDTO(session, attendeeCount);
+        return SessionConverter.toOrganizerSessionInfoDTO(session, attendeeCount);
     }
 
     @Override
@@ -111,28 +106,11 @@ public class OrganizerEventServiceImpl implements OrganizerEventService {
         BigInteger priceWei = exchangeService.convertKrwToWei(BigDecimal.valueOf(request.getPriceKrw()));
 
         Event event = toEvent(request, user, bannerUrl, posterUrl, priceWei);
-
-        event.setEventStatus(EventStatus.APPLY_NOT_OPENED);
         eventRepository.save(event);
+        user.addEvent(event);
 
         photoCardService.createPhotoCards(event, photocardList);
-
-        for (LocalDate date = request.getStartDate(); !date.isAfter(request.getEndDate()); date = date.plusDays(1)) {
-            Session session = Session.builder()
-                    .event(event)
-                    .date(date)
-                    .isDrawn(false)
-                    .metadataUploaded(false)
-                    .isBuyable(false)
-                    .build();
-
-            event.addSession(session);
-
-            sessionRepository.save(session);
-            schedulingService.scheduleSessionJob(session, ClosePaymentJob.class);
-        }
-
-        user.addEvent(event);
+        sessionService.createSessions(event);
 
         schedulingService.scheduleEventJob(event, OpenApplyJob.class);
         event.setTxHash(dketNFTService.recordEventOnChain(event));
@@ -142,18 +120,20 @@ public class OrganizerEventServiceImpl implements OrganizerEventService {
                 .build();
     }
 
+    private void validateOrganizer(Event event, User user) {
+        if (!event.getOrganizer().getId().equals(user.getId())) {
+            throw new CustomException(ErrorStatus.EVENT_ORGANIZER_MISMATCH);
+        }
+    }
+
     private void validateSchedule(EventUploadDTO request) {
-        if (request.getApplyStart().isBefore(LocalDateTime.now()))
-            throw new CustomException(ErrorStatus.EVENT_INVALID_SCHEDULE);
-
-        if (!request.getApplyEnd().isAfter(request.getApplyStart()))
-            throw new CustomException(ErrorStatus.EVENT_INVALID_SCHEDULE);
-
-        if (request.getApplyEnd().plusDays(Constants.PAYMENT_DEADLINE).withHour(0).withMinute(0)
+        if ((request.getApplyStart().isBefore(LocalDateTime.now()))
+            || (!request.getApplyEnd().isAfter(request.getApplyStart()))
+            || (request.getApplyEnd().plusDays(Constants.PAYMENT_DEADLINE).withHour(0).withMinute(0)
                 .isAfter(LocalDateTime.of(request.getStartDate(), LocalTime.of(0, 0))))
+            || (request.getEndDate().isBefore(request.getStartDate())))
+        {
             throw new CustomException(ErrorStatus.EVENT_INVALID_SCHEDULE);
-
-        if (request.getEndDate().isBefore(request.getStartDate()))
-            throw new CustomException(ErrorStatus.EVENT_INVALID_SCHEDULE);
+        }
     }
 }
