+++ src/main/java/com/example/demo/domain/event/service/impl/SessionServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.demo.domain.event.service.impl;
 
+import com.example.demo.domain.apply.entity.Apply;
 import com.example.demo.domain.apply.enums.ApplyStatus;
 import com.example.demo.domain.apply.repository.ApplyRepository;
 import com.example.demo.domain.event.entity.Event;
@@ -7,9 +8,13 @@ import com.example.demo.domain.event.entity.Session;
 import com.example.demo.domain.event.enums.EventStatus;
 import com.example.demo.domain.event.repository.SessionRepository;
 import com.example.demo.domain.event.service.SessionService;
+import com.example.demo.domain.event.dto.response.PriceWeiDTO;
+import com.example.demo.domain.user.entity.User;
+import com.example.demo.domain.user.service.UserService;
 import com.example.demo.global.event.ReadyToMintEvent;
 import com.example.demo.global.infra.scheduling.SchedulingService;
 import com.example.demo.global.infra.scheduling.jobs.event.OpenPublicJob;
+import com.example.demo.global.infra.scheduling.jobs.session.ClosePaymentJob;
 import com.example.demo.global.response.exception.CustomException;
 import com.example.demo.global.response.status.ErrorStatus;
 import lombok.RequiredArgsConstructor;
@@ -17,6 +22,8 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -28,6 +35,7 @@ public class SessionServiceImpl implements SessionService {
     private final ApplyRepository applyRepository;
     private final SchedulingService schedulingService;
     private final ApplicationEventPublisher eventPublisher;
+    private final UserService userService;
 
     @Override
     @Transactional
@@ -59,7 +67,72 @@ public class SessionServiceImpl implements SessionService {
 
         session.setIsDrawn();
 
-        if(session.getMetadataUploaded())
+        if(session.getMetadataUploaded()) {
             eventPublisher.publishEvent(new ReadyToMintEvent(session.getId()));
+        }
+    }
+
+    @Override
+    @Transactional
+    public void createSessions(Event event) {
+        List<Session> sessionList = new ArrayList<>();
+        for (LocalDate date = event.getStartDate(); !date.isAfter(event.getEndDate()); date = date.plusDays(1)) {
+            Session session = Session.builder()
+                    .event(event)
+                    .date(date)
+                    .isDrawn(false)
+                    .metadataUploaded(false)
+                    .isBuyable(false)
+                    .build();
+
+            event.addSession(session);
+
+            sessionList.add(session);
+            schedulingService.scheduleSessionJob(session, ClosePaymentJob.class);
+        }
+
+        sessionRepository.saveAll(sessionList);
+    }
+
+    @Override
+    public PriceWeiDTO getPriceWei(Long sessionId) {
+        User user = userService.getCurrentUser();
+        Session session = sessionRepository.findById(sessionId)
+                .orElseThrow(() -> new CustomException(ErrorStatus.SESSION_NOT_FOUND));
+
+        validateBuyer(session, user);
+
+        return PriceWeiDTO.builder()
+                .priceWei(session.getEvent().getPriceWei())
+                .sessionId(sessionId)
+                .build();
+    }
+
+    private void validateBuyer(Session session, User user) {
+        Event event = session.getEvent();
+
+        if (!user.isEligibleFor(event.getAgeLimit()))
+            throw new CustomException(ErrorStatus.TICKET_INVALID_BUYER);
+
+        if (!session.getIsBuyable())
+            throw new CustomException(ErrorStatus.SESSION_CANNOT_BUY);
+
+        switch (event.getEventStatus()) {
+            case APPLY_CLOSED:
+                Apply apply = applyRepository.findBySessionIdAndUserId(session.getId(), user.getId())
+                        .orElseThrow(() -> new CustomException(ErrorStatus.APPLY_NOT_FOUND));
+
+                if (apply.getApplyStatus() != ApplyStatus.SELECTED)
+                    throw new CustomException(ErrorStatus.TICKET_INVALID_BUYER);
+
+                break;
+
+            case TICKETED:
+            case IN_PROGRESS:
+                break;
+
+            default:
+                throw new CustomException(ErrorStatus.SESSION_CANNOT_BUY);
+        }
     }
 }
