+++ src/main/java/com/example/demo/domain/main/service/BuyerHomeService.java
@@ -2,7 +2,6 @@ package com.example.demo.domain.main.service;
 
 import com.example.demo.domain.main.dto.BuyerHomeResponseDTO;
 import com.example.demo.domain.main.dto.EventCardListDTO;
-import org.springframework.data.domain.Pageable;
 
 public interface BuyerHomeService {
