+++ src/main/java/com/example/demo/domain/ticket/controller/TicketController.java
@@ -1,6 +1,7 @@
 package com.example.demo.domain.ticket.controller;
 
-import com.example.demo.domain.ticket.dto.PriceWeiDTO;
+import com.example.demo.domain.event.dto.response.PriceWeiDTO;
+import com.example.demo.domain.event.service.SessionService;
 import com.example.demo.domain.ticket.dto.TicketDetailDTO;
 import com.example.demo.domain.ticket.service.TicketService;
 import com.example.demo.global.response.ApiResponse;
@@ -18,6 +19,7 @@ import static com.example.demo.global.response.status.SuccessStatus._OK;
 public class TicketController {
 
     private final TicketService ticketService;
+    private final SessionService sessionService;
 
     @Operation(summary = "티켓 조회")
     @GetMapping("")
@@ -46,6 +48,6 @@ public class TicketController {
     @Operation(summary = "티켓 가격 확인")
     @GetMapping("buyer/{sessionId}")
     public ApiResponse<PriceWeiDTO> getPriceWei(@PathVariable("sessionId") Long sessionId) {
-        return ApiResponse.onSuccess(_OK, ticketService.getPriceWei(sessionId));
+        return ApiResponse.onSuccess(_OK, sessionService.getPriceWei(sessionId));
     }
 }
