+++ src/main/java/com/example/demo/global/infra/blockchain/service/impl/DketNFTServiceImpl.java
@@ -1,0 +1,338 @@
+package com.example.demo.global.infra.blockchain.service.impl;
+
+import com.example.demo.domain.event.entity.Event;
+import com.example.demo.domain.event.entity.Session;
+import com.example.demo.domain.event.repository.SessionRepository;
+import com.example.demo.domain.event.service.SessionService;
+import com.example.demo.domain.metadata.service.MetadataService;
+import com.example.demo.domain.ticket.service.TicketService;
+import com.example.demo.global.event.ReadyToMintEvent;
+import com.example.demo.global.infra.blockchain.contracts.DketNFT;
+import com.example.demo.global.infra.blockchain.service.DketNFTService;
+import com.example.demo.global.response.exception.CustomException;
+import com.example.demo.global.response.status.ErrorStatus;
+import jakarta.annotation.PostConstruct;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.event.EventListener;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.web3j.abi.FunctionEncoder;
+import org.web3j.abi.datatypes.DynamicArray;
+import org.web3j.abi.datatypes.Function;
+import org.web3j.abi.datatypes.Type;
+import org.web3j.abi.datatypes.Utf8String;
+import org.web3j.abi.datatypes.generated.Uint256;
+import org.web3j.crypto.Credentials;
+import org.web3j.protocol.Web3j;
+import org.web3j.protocol.core.DefaultBlockParameterName;
+import org.web3j.protocol.core.Response;
+import org.web3j.protocol.core.methods.request.Transaction;
+import org.web3j.protocol.core.methods.response.EthEstimateGas;
+import org.web3j.protocol.core.methods.response.EthSendTransaction;
+import org.web3j.tx.RawTransactionManager;
+import org.web3j.tx.exceptions.ContractCallException;
+import org.web3j.tx.gas.DefaultGasProvider;
+
+import java.math.BigInteger;
+import java.util.*;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class DketNFTServiceImpl implements DketNFTService {
+
+    private final Web3j web3j;
+    private final Credentials credentials;
+    private final MetadataService metadataService;
+    private final SessionRepository sessionRepository;
+    private final TicketService ticketService;
+    private final SessionService sessionService;
+    private final ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1);
+
+
+    @Value("${web3.contract-address}")
+    private String contractAddress;
+
+    private DketNFT dketNFT;
+
+    @PostConstruct
+    public void init() {
+        dketNFT = DketNFT.load(
+                contractAddress,
+                web3j,
+                credentials,
+                new DefaultGasProvider()
+        );
+
+        listenToRandomFulfilled();
+        listenToWinnersDrawn();
+        listenToSetDrawn();
+        listenToSessionMinted();
+        listenToPaymentTransferred();
+    }
+
+    @Override
+    public String recordEventOnChain(Event event) {
+        try {
+            var tx = dketNFT.createEvent(
+                    BigInteger.valueOf(event.getId()),
+                    event.getOrganizer().getWalletAddress(),
+                    event.getTitle(),
+                    BigInteger.valueOf(event.getCapacity()),
+                    event.getPriceWei()
+            ).send();
+
+            return tx.getTransactionHash();
+        } catch (Exception e) {
+            log.error("Event [{}] 온체인 기록 실패", event.getId(), e);
+            throw new CustomException(ErrorStatus.BLOCKCHAIN_TRANSACTION_FAILED);
+        }
+    }
+
+    @Override
+    @Transactional
+    public void recordAllSessionsOnChain(Event event) {
+        List<Session> sessions = event.getSessions();
+
+        Map<Session, String> txMap = sessions.parallelStream()
+            .collect(Collectors.toMap(
+                session -> session,
+                this::recordSessionOnChain
+            ));
+
+        txMap.forEach((session, txHash) -> session.setTxHash(txHash));
+    }
+
+    private String recordSessionOnChain(Session session) {
+        try {
+            List<String> applications = session.getApplyList().stream()
+                    .map(apply -> apply.getUser().getWalletAddress())
+                    .collect(Collectors.toList());
+
+            var tx = dketNFT.createSession(
+                    BigInteger.valueOf(session.getEvent().getId()),
+                    BigInteger.valueOf(session.getId()),
+                    applications
+            ).send();
+
+            return tx.getTransactionHash();
+        } catch (Exception e) {
+            log.error("Session [{}] 온체인 기록 실패", session.getId(), e);
+            throw new CustomException(ErrorStatus.BLOCKCHAIN_TRANSACTION_FAILED);
+        }
+    }
+
+    @Override
+    public void openPublicSaleOnChain(Event event) {
+        try {
+            dketNFT.openPublicSale(BigInteger.valueOf(event.getId())).send();
+        } catch (Exception e) {
+            log.error("Event [{}] 온체인 선착순 판매 전환 실패", event.getId(), e);
+            throw new CustomException(ErrorStatus.BLOCKCHAIN_TRANSACTION_FAILED);
+        }
+    }
+
+    private void listenToRandomFulfilled() {
+        dketNFT.randomFulfilledEventFlowable(DefaultBlockParameterName.LATEST, DefaultBlockParameterName.LATEST)
+            .subscribe(
+                event -> {
+                    BigInteger sessionId = event.sessionId;
+                    BigInteger randomWord = event.randomWord;
+
+                    List<Long> metadataIds = metadataService.createMetadata(sessionId, randomWord);
+                    metadataService.uploadAllMetadataAsync(metadataIds);
+
+                    scheduler.schedule(() -> {
+                        try {
+                            drawSession(sessionId);
+                        } catch (Exception e) {
+                            log.error("Session [{}] 추첨 지연 실패", sessionId, e);
+                        }
+                    }, 3, TimeUnit.SECONDS);
+                },
+                error -> {
+                    log.error("randomFulfilled 이벤트 수신 중 에러 발생", error);
+                }
+            );
+    }
+
+    @Transactional
+    protected void drawSession(BigInteger sessionId) {
+        List<Type> input = List.of(new Uint256(sessionId));
+        sendTransaction("drawSession", input);
+    }
+
+    private void listenToWinnersDrawn() {
+        dketNFT.winnersDrawnEventFlowable(DefaultBlockParameterName.LATEST, DefaultBlockParameterName.LATEST)
+            .subscribe(
+                event -> {
+                    BigInteger sessionId = event.sessionId;
+                    List<String> winners = event.winners;
+
+                    if (winners == null || winners.isEmpty()) {
+                        throw new CustomException(ErrorStatus.SESSION_DRAW_FAILED);
+                    }
+
+                    sessionService.saveWinners(sessionId.longValue(), winners);
+                },
+                error -> {
+                    log.error("winnersDrawn 이벤트 수신 중 예외 발생", error);
+                }
+            );
+    }
+
+    private void listenToSetDrawn() {
+        dketNFT.setDrawnEventFlowable(DefaultBlockParameterName.LATEST, DefaultBlockParameterName.LATEST)
+            .subscribe(
+                event -> {
+                    Long sessionId = event.sessionId.longValue();
+                    sessionService.completeDraw(sessionId);
+                },
+                error -> {
+                    log.error("setDrawn 이벤트 수신 중 예외 발생", error);
+                }
+            );
+    }
+
+    @EventListener
+    public void handleMetadataUploaded(ReadyToMintEvent event) {
+        Session session = sessionRepository.findById(event.getSessionId())
+                .orElseThrow(() -> new CustomException(ErrorStatus.SESSION_NOT_FOUND));
+
+        mintSessionTicket(session);
+    }
+
+    private void mintSessionTicket(Session session) {
+        List<String> metadataUris = session.getMetadataList().stream()
+                .map(metadata -> "ipfs://" + metadata.getCid())
+                .collect(Collectors.toList());
+
+        BigInteger sessionId = BigInteger.valueOf(session.getId());
+
+        List<Type> input = Arrays.asList(
+                new Uint256(sessionId),
+                new DynamicArray<>(
+                    Utf8String.class,
+                    metadataUris.stream()
+                        .map(Utf8String::new)
+                        .collect(Collectors.toList())
+                )
+        );
+        sendTransaction("mintSessionTicket", input);
+    }
+
+    private void listenToSessionMinted() {
+        dketNFT.sessionMintedEventFlowable(DefaultBlockParameterName.LATEST, DefaultBlockParameterName.LATEST)
+            .subscribe(
+                event -> {
+                    List<BigInteger> tokenIds = event.tokenIds;
+
+                    if (tokenIds == null || tokenIds.isEmpty()) {
+                        throw new CustomException(ErrorStatus.SESSION_MINTING_FAILED);
+                    }
+
+                    registerMintedTickets(tokenIds);
+                },
+                error -> {
+                    log.error("sessionMinted 이벤트 수신 중 예외 발생", error);
+                }
+            );
+    }
+
+    private void registerMintedTickets(List<BigInteger> tokenIds) {
+        List<String> cidList = new ArrayList<>();
+
+        for (BigInteger tokenId : tokenIds) {
+            cidList.add(getTokenUri(tokenId).replace("ipfs://", ""));
+        }
+
+        ticketService.batchRegisterTicket(tokenIds, cidList);
+    }
+
+    private void listenToPaymentTransferred() {
+        dketNFT.paymentTransferredEventFlowable(DefaultBlockParameterName.LATEST, DefaultBlockParameterName.LATEST)
+            .subscribe(
+                event -> {
+                    String buyer = event.to;
+                    Long sessionId = event.sessionId.longValue();
+                    Long tokenId = event.tokenId.longValue();
+
+                    ticketService.completeTicket(buyer, sessionId, tokenId);
+                },
+                error -> {
+                    log.error("paymentTransferred 이벤트 수신 중 예외 발생", error);
+                }
+            );
+    }
+
+    public void sendTransaction(String functionName, List<Type> inputParams) {
+        try {
+            Function function = new Function(functionName, inputParams, Collections.emptyList());
+            String encodedFunction = FunctionEncoder.encode(function);
+            BigInteger gasLimit = estimateGas(encodedFunction);
+            BigInteger gasPrice = web3j.ethGasPrice().send().getGasPrice();
+
+            RawTransactionManager txManager = new RawTransactionManager(web3j, credentials);
+            EthSendTransaction response = txManager.sendTransaction(
+                    gasPrice,
+                    gasLimit,
+                    dketNFT.getContractAddress(),
+                    encodedFunction,
+                    BigInteger.ZERO
+            );
+
+            if (response.hasError()) {
+                log.error("{} 트랜잭션 실패: {}", functionName, response.getError().getMessage());
+                throw new CustomException(ErrorStatus.BLOCKCHAIN_TRANSACTION_FAILED);
+            }
+        } catch (Exception e) {
+            log.error("{} 실패", functionName, e);
+            throw new CustomException(ErrorStatus.BLOCKCHAIN_TRANSACTION_FAILED);
+        }
+    }
+
+    private BigInteger estimateGas(String encodedFunction) {
+        try {
+            Transaction ethCallTransaction = Transaction.createEthCallTransaction(
+                    credentials.getAddress(),
+                    dketNFT.getContractAddress(),
+                    encodedFunction
+            );
+
+            EthEstimateGas ethEstimateGas = web3j.ethEstimateGas(ethCallTransaction).send();
+
+            if (ethEstimateGas.hasError()) {
+                Response.Error error = ethEstimateGas.getError();
+                log.error("Gas 추정 실패 - Code: {}, Message: {}, Data: {}",
+                        error.getCode(), error.getMessage(), error.getData());
+                throw new CustomException(ErrorStatus.BLOCKCHAIN_ESTIMATE_GAS_FAILED);
+            }
+
+            return ethEstimateGas.getAmountUsed().multiply(BigInteger.valueOf(120)).divide(BigInteger.valueOf(100));
+
+        } catch (Exception e) {
+            log.error("Gas 추정 중 예외 발생", e);
+            throw new CustomException(ErrorStatus.BLOCKCHAIN_ESTIMATE_GAS_FAILED);
+        }
+    }
+
+    private String getTokenUri(BigInteger tokenId) {
+        try {
+            return dketNFT.tokenURI(tokenId).send();
+        } catch (ContractCallException e) {
+            log.error("Token [{}] URI 호출 실패: {}", tokenId, e.getMessage(), e);
+            throw new CustomException(ErrorStatus.TOKEN_INVALID);
+        } catch (Exception e) {
+            log.error("Token [{}] TokenUri 트랜잭션 실패: {}", tokenId, e.getMessage(), e);
+            throw new CustomException(ErrorStatus.BLOCKCHAIN_ESTIMATE_GAS_FAILED);
+        }
+    }
+}
